Extract to .h file
